requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==2.1.1
coverage==6.5.0
Django==4.1.5
django-environ==0.9.0
django-rest-passwordreset==1.3.0
djangorestframework==3.14.0
environ==1.0
idna==3.4
psycopg2-binary==2.9.5
pytz==2022.7
PyYAML==6.0
requests==2.28.1
sqlparse==0.4.3
ujson==5.7.0
urllib3==1.26.13
gunicorn==20.1.0
drf_yasg
celery==5.2.7
redis
django-allauth